karaousel_service/requirements.txt
aiosmtplib==2.0.0
anyio==3.6.2
bcrypt==4.0.1
black==22.12.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==3.4.8
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.4
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.8.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
pycparser==2.21
pydantic==1.10.3
PyJWT==1.7.1
pymongo==4.3.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.7.0
ujson==5.6.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
#ariadne==0.21
strawberry-graphql[fastapi]
